Auth with Identity server

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -4,6 +4,7 @@ using AuctionService.Entities;
 using AutoMapper;
 using BuildingBlocks.Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -69,12 +70,13 @@ namespace AuctionService.Controllers
             return _mapper.Map<AuctionDto>(auction);
         }
 
+        [Authorize]
         [HttpPost]
         public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto auctionDto)
         {
             var auction = _mapper.Map<Auction>(auctionDto);
-            // TODO: add current user as a seller
-            auction.Seller = "Ahmad";
+
+            auction.Seller = User.Identity!.Name!;
 
             _context.Auctions.Add(auction);
 
@@ -94,6 +96,7 @@ namespace AuctionService.Controllers
             );
         }
 
+        [Authorize]
         [HttpPut("{id}")]
         public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDto auctionDto)
         {
@@ -101,9 +104,11 @@ namespace AuctionService.Controllers
                 .Auctions.Include(a => a.Item)
                 .FirstOrDefaultAsync(a => a.Id == id);
 
-            // TODO: check the current user is the seller
             if (auction is null)
                 return BadRequest();
+
+            if (auction.Seller != User.Identity!.Name)
+                return Forbid();
 
             auction.Item.Make = auctionDto.Make ?? auction.Item.Make;
             auction.Item.Model = auctionDto.Model ?? auction.Item.Model;
@@ -123,15 +128,17 @@ namespace AuctionService.Controllers
             return Ok();
         }
 
+        [Authorize]
         [HttpDelete("{id}")]
         public async Task<ActionResult> DeleteAuction(Guid id)
         {
             var auction = await _context.Auctions.FindAsync(id);
 
-            // TODO: check the current user is the seller
-
             if (auction is null)
                 return BadRequest();
+
+            if (auction.Seller != User.Identity!.Name)
+                return Forbid();
 
             _context.Auctions.Remove(auction);
